+++ e2e/__tests__/jestChangedFiles.test.ts
@@ -32,7 +32,7 @@ if (!hgIsInstalled) {
   console.warn('Mercurial (hg) is not installed - skipping some tests');
 }
 
-testIfHg('gets hg SCM roots and dedups them', async () => {
+testIfHg('gets hg SCM roots and dedupes them', async () => {
   writeFiles(DIR, {
     'first-repo/file1.txt': 'file1',
     'first-repo/nested-dir/file2.txt': 'file2',
@@ -67,7 +67,7 @@ testIfHg('gets hg SCM roots and dedups them', async () => {
   expect(hgRepos[1]).toMatch(/\/jest-changed-files-test-dir\/second-repo\/?$/);
 });
 
-test('gets git SCM roots and dedups them', async () => {
+test('gets git SCM roots and dedupes them', async () => {
   writeFiles(DIR, {
     'first-repo/file1.txt': 'file1',
     'first-repo/nested-dir/file2.txt': 'file2',
@@ -101,7 +101,7 @@ test('gets git SCM roots and dedups them', async () => {
   expect(gitRepos[1]).toMatch(/\/jest-changed-files-test-dir\/second-repo\/?$/);
 });
 
-testIfHg('gets mixed git and hg SCM roots and dedups them', async () => {
+testIfHg('gets mixed git and hg SCM roots and dedupes them', async () => {
   writeFiles(DIR, {
     'first-repo/file1.txt': 'file1',
     'first-repo/nested-dir/file2.txt': 'file2',
@@ -179,6 +179,16 @@ test('gets changed files for git', async () => {
     'file1.txt': 'modified file1',
   });
 
+  ({changedFiles: files} = await getChangedFilesForRoots(roots, {}));
+  expect(
+    Array.from(files)
+      .map(filePath => path.basename(filePath))
+      .sort(),
+  ).toEqual(['file1.txt']);
+
+  run(`${GIT} add -A`, DIR);
+
+  // staged files should be included
   ({changedFiles: files} = await getChangedFilesForRoots(roots, {}));
   expect(
     Array.from(files)